requirements.txt
openai>=1.0.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
groq>=0.4.0
anthropic>=0.7.0
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
